Update to evaluation function
Update to GetDataFunctions:
You can either save the predictions to a file and then evaluate that file, or now you can just put in the variables. The save to file will be useful later for the preliminary report and the leaderboard.

Also added optional seed parameter to the split train and test functions.