My final tweaks
Updates to Experimental Functions:
Small fix to the series_metric function to allow rows within the window to have an average, even if it's not the full past/future.

Created MITL file.

Updates to Vega_Final_Project:
-Added the data plot showing the DIS instrument data compared to the other data.
-Changed the qqplot to be the same theme as later CV plot.
-Added a test.data set for use
 in prediction for the MITL predictions file.
-Rearranged the document outline for better flow.
-Added evaluation criteria section.
-Changed table to only include test data; this makes it more interpretable.
-Reworded some things.
-Changed validation set size to 10% instead of 5%.
-Added a Data Exploration section.

Moved: Several files have been moved out of the main folder so that it is easier to tell which files need to be included in the final report zip.

Noted: The tree depth and number of trees variables are moved up higher in the document so they can be changed more easily. Also, there are a significant number of libraries included. I would have prefixed them with namespaces, however if I do this, and you don't ahve the package installed, you will have gone very far into the knitting process before you get an error. (The knitting process is very long for this document.) To avoid this, I've tried to include all the libraries needed at the top of the document.